refactor(protocol): split wire compressed_token zstd_codec into submodules
